+++ core/libdeno.rs
@@ -26,9 +26,6 @@ use std::slice;
 
 pub type OpId = u32;
 
-#[allow(non_camel_case_types)]
-pub type isolate = DenoIsolate;
-
 pub struct ModuleInfo {
   pub main: bool,
   pub name: String,
@@ -36,456 +33,7 @@ pub struct ModuleInfo {
   pub import_specifiers: Vec<String>,
 }
 
-pub struct DenoIsolate {
-  pub isolate_: Option<v8::OwnedIsolate>,
-  pub last_exception_: Option<String>,
-  pub last_exception_handle_: v8::Global<v8::Value>,
-  pub context_: v8::Global<v8::Context>,
-  pub mods_: HashMap<deno_mod, ModuleInfo>,
-  pub mods_by_name_: HashMap<String, deno_mod>,
-  pub locker_: Option<v8::Locker>,
-  pub shared_: deno_buf,
-  pub shared_ab_: v8::Global<v8::SharedArrayBuffer>,
-  pub resolve_cb_: Option<deno_resolve_cb>,
-  pub recv_: v8::Global<v8::Function>,
-  pub current_args_: *const v8::FunctionCallbackInfo,
-  pub recv_cb_: deno_recv_cb,
-  pub snapshot_creator_: Option<v8::SnapshotCreator>,
-  pub has_snapshotted_: bool,
-  pub snapshot_: Option<SnapshotConfig>,
-  pub next_dyn_import_id_: deno_dyn_import_id,
-  pub dyn_import_cb_: deno_dyn_import_cb,
-  pub dyn_import_map_:
-    HashMap<deno_dyn_import_id, v8::Global<v8::PromiseResolver>>,
-  pub pending_promise_map_: HashMap<i32, v8::Global<v8::Value>>,
-  // Used in deno_mod_instantiate
-  pub resolve_context_: *mut c_void,
-  // Used in deno_mod_evaluate
-  pub core_isolate_: *mut c_void,
-  /*
-  void* global_import_buf_ptr_;
-  v8::Persistent<v8::ArrayBuffer> global_import_buf_;
-  */
-}
-
-impl Drop for DenoIsolate {
-  fn drop(&mut self) {
-    // TODO Too much boiler plate.
-    // <Boilerplate>
-    let mut isolate = self.isolate_.take().unwrap();
-    {
-      let mut locker = v8::Locker::new(&isolate);
-      let mut hs = v8::HandleScope::new(&mut locker);
-      let scope = hs.enter();
-      // </Boilerplate>
-      self.context_.reset(scope);
-      self.shared_ab_.reset(scope);
-      self.last_exception_handle_.reset(scope);
-      self.recv_.reset(scope);
-      for (key, module) in self.mods_.iter_mut() {
-        module.handle.reset(scope);
-      }
-      for (key, handle) in self.dyn_import_map_.iter_mut() {
-        handle.reset(scope);
-      }
-      for (key, handle) in self.pending_promise_map_.iter_mut() {
-        handle.reset(scope);
-      }
-    }
-    if let Some(locker_) = self.locker_.take() {
-      drop(locker_);
-    }
-    if let Some(creator) = self.snapshot_creator_.take() {
-      // TODO(ry) V8 has a strange assert which prevents a SnapshotCreator from
-      // being deallocated if it hasn't created a snapshot yet.
-      // https://github.com/v8/v8/blob/73212783fbd534fac76cc4b66aac899c13f71fc8/src/api.cc#L603
-      // If that assert is removed, this if guard could be removed.
-      // WARNING: There may be false positive LSAN errors here.
-      std::mem::forget(isolate);
-      if self.has_snapshotted_ {
-        drop(creator);
-      }
-    } else {
-      drop(isolate);
-    }
-  }
-}
-
-impl DenoIsolate {
-  pub fn new(config: deno_config) -> Self {
-    Self {
-      isolate_: None,
-      last_exception_: None,
-      last_exception_handle_: v8::Global::<v8::Value>::new(),
-      context_: v8::Global::<v8::Context>::new(),
-      mods_: HashMap::new(),
-      mods_by_name_: HashMap::new(),
-      pending_promise_map_: HashMap::new(),
-      locker_: None,
-      shared_: config.shared,
-      shared_ab_: v8::Global::<v8::SharedArrayBuffer>::new(),
-      resolve_cb_: None,
-      recv_: v8::Global::<v8::Function>::new(),
-      current_args_: std::ptr::null(),
-      recv_cb_: config.recv_cb,
-      snapshot_creator_: None,
-      snapshot_: config.load_snapshot,
-      has_snapshotted_: false,
-      next_dyn_import_id_: 0,
-      dyn_import_cb_: config.dyn_import_cb,
-      dyn_import_map_: HashMap::new(),
-      resolve_context_: std::ptr::null_mut(),
-      core_isolate_: std::ptr::null_mut(),
-    }
-  }
-
-  pub fn add_isolate(&mut self, mut isolate: v8::OwnedIsolate) {
-    isolate.set_capture_stack_trace_for_uncaught_exceptions(true, 10);
-    isolate.set_promise_reject_callback(bindings::promise_reject_callback);
-    isolate.add_message_listener(bindings::message_callback);
-    isolate.set_host_initialize_import_meta_object_callback(
-      bindings::host_initialize_import_meta_object_callback,
-    );
-    isolate.set_host_import_module_dynamically_callback(
-      bindings::host_import_module_dynamically_callback,
-    );
-    let self_ptr: *mut Self = self;
-    unsafe { isolate.set_data(0, self_ptr as *mut c_void) };
-    self.isolate_ = Some(isolate);
-  }
-
-  pub fn register_module(
-    &mut self,
-    main: bool,
-    name: &str,
-    source: &str,
-  ) -> deno_mod {
-    let isolate = self.isolate_.as_ref().unwrap();
-    let mut locker = v8::Locker::new(&isolate);
-
-    let mut hs = v8::HandleScope::new(&mut locker);
-    let scope = hs.enter();
-    let mut context = v8::Context::new(scope);
-    context.enter();
-
-    let name_str = v8::String::new(scope, name).unwrap();
-    let source_str = v8::String::new(scope, source).unwrap();
-
-    let origin = module_origin(scope, name_str);
-    let source = v8::script_compiler::Source::new(source_str, &origin);
-
-    let mut try_catch = v8::TryCatch::new(scope);
-    let tc = try_catch.enter();
-
-    let mut maybe_module =
-      v8::script_compiler::compile_module(&isolate, source);
-
-    if tc.has_caught() {
-      assert!(maybe_module.is_none());
-      self.handle_exception(scope, context, tc.exception().unwrap());
-      context.exit();
-      return 0;
-    }
-    let module = maybe_module.unwrap();
-    let id = module.get_identity_hash();
-
-    let mut import_specifiers: Vec<String> = vec![];
-    for i in 0..module.get_module_requests_length() {
-      let specifier = module.get_module_request(i);
-      import_specifiers.push(specifier.to_rust_string_lossy(scope));
-    }
-
-    let mut handle = v8::Global::<v8::Module>::new();
-    handle.set(scope, module);
-    self.mods_.insert(
-      id,
-      ModuleInfo {
-        main,
-        name: name.to_string(),
-        import_specifiers,
-        handle,
-      },
-    );
-    self.mods_by_name_.insert(name.to_string(), id);
-    context.exit();
-    id
-  }
-
-  pub fn get_module_info(&self, id: deno_mod) -> Option<&ModuleInfo> {
-    if id == 0 {
-      return None;
-    }
-    self.mods_.get(&id)
-  }
-
-  fn execute<'a>(
-    &mut self,
-    s: &mut impl v8::ToLocal<'a>,
-    mut context: v8::Local<'a, v8::Context>,
-    js_filename: &str,
-    js_source: &str,
-  ) -> bool {
-    let mut hs = v8::HandleScope::new(s);
-    let s = hs.enter();
-    let source = v8::String::new(s, js_source).unwrap();
-    let name = v8::String::new(s, js_filename).unwrap();
-    let mut try_catch = v8::TryCatch::new(s);
-    let tc = try_catch.enter();
-    let origin = script_origin(s, name);
-    let mut script =
-      v8::Script::compile(s, context, source, Some(&origin)).unwrap();
-    let result = script.run(s, context);
-    if result.is_none() {
-      assert!(tc.has_caught());
-      let exception = tc.exception().unwrap();
-      self.handle_exception(s, context, exception);
-      false
-    } else {
-      true
-    }
-  }
-
-  pub fn handle_exception<'a>(
-    &mut self,
-    s: &mut impl v8::ToLocal<'a>,
-    mut context: v8::Local<'a, v8::Context>,
-    exception: v8::Local<'a, v8::Value>,
-  ) {
-    let isolate = context.get_isolate();
-    // TerminateExecution was called
-    if isolate.is_execution_terminating() {
-      // cancel exception termination so that the exception can be created
-      isolate.cancel_terminate_execution();
-
-      // maybe make a new exception object
-      let exception = if (exception.is_null_or_undefined()) {
-        let exception_str = v8::String::new(s, "execution terminated").unwrap();
-        isolate.enter();
-        let e = v8::error(s, exception_str);
-        isolate.exit();
-        e
-      } else {
-        exception
-      };
-
-      // handle the exception as if it is a regular exception
-      self.handle_exception(s, context, exception);
-
-      // re-enable exception termination
-      context.get_isolate().terminate_execution();
-      return;
-    }
-
-    let json_str = self.encode_exception_as_json(s, context, exception);
-    self.last_exception_ = Some(json_str);
-    self.last_exception_handle_.set(s, exception);
-  }
-
-  pub fn encode_exception_as_json<'a>(
-    &mut self,
-    s: &mut impl v8::ToLocal<'a>,
-    mut context: v8::Local<'a, v8::Context>,
-    exception: v8::Local<'a, v8::Value>,
-  ) -> String {
-    let message = v8::create_message(s, exception);
-    self.encode_message_as_json(s, context, message)
-  }
-
-  pub fn encode_message_as_json<'a>(
-    &mut self,
-    s: &mut impl v8::ToLocal<'a>,
-    mut context: v8::Local<v8::Context>,
-    message: v8::Local<v8::Message>,
-  ) -> String {
-    let json_obj = self.encode_message_as_object(s, context, message);
-    let json_string = v8::json::stringify(context, json_obj.into()).unwrap();
-    json_string.to_rust_string_lossy(s)
-  }
-
-  fn encode_message_as_object<'a>(
-    &mut self,
-    s: &mut impl v8::ToLocal<'a>,
-    mut context: v8::Local<v8::Context>,
-    message: v8::Local<v8::Message>,
-  ) -> v8::Local<'a, v8::Object> {
-    let json_obj = v8::Object::new(s);
-
-    let exception_str = message.get(s);
-    json_obj.set(
-      context,
-      v8::String::new(s, "message").unwrap().into(),
-      exception_str.into(),
-    );
-
-    let script_resource_name = message
-      .get_script_resource_name(s)
-      .expect("Missing ScriptResourceName");
-    json_obj.set(
-      context,
-      v8::String::new(s, "scriptResourceName").unwrap().into(),
-      script_resource_name,
-    );
-
-    let source_line = message
-      .get_source_line(s, context)
-      .expect("Missing SourceLine");
-    json_obj.set(
-      context,
-      v8::String::new(s, "sourceLine").unwrap().into(),
-      source_line.into(),
-    );
-
-    let line_number = message
-      .get_line_number(context)
-      .expect("Missing LineNumber");
-    json_obj.set(
-      context,
-      v8::String::new(s, "lineNumber").unwrap().into(),
-      v8::Integer::new(s, line_number as i32).into(),
-    );
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "startPosition").unwrap().into(),
-      v8::Integer::new(s, message.get_start_position() as i32).into(),
-    );
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "endPosition").unwrap().into(),
-      v8::Integer::new(s, message.get_end_position() as i32).into(),
-    );
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "errorLevel").unwrap().into(),
-      v8::Integer::new(s, message.error_level() as i32).into(),
-    );
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "startColumn").unwrap().into(),
-      v8::Integer::new(s, message.get_start_column() as i32).into(),
-    );
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "endColumn").unwrap().into(),
-      v8::Integer::new(s, message.get_end_column() as i32).into(),
-    );
-
-    let is_shared_cross_origin =
-      v8::Boolean::new(s, message.is_shared_cross_origin());
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "isSharedCrossOrigin").unwrap().into(),
-      is_shared_cross_origin.into(),
-    );
-
-    let is_opaque = v8::Boolean::new(s, message.is_opaque());
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "isOpaque").unwrap().into(),
-      is_opaque.into(),
-    );
-
-    let frames = if let Some(stack_trace) = message.get_stack_trace(s) {
-      let count = stack_trace.get_frame_count() as i32;
-      let frames = v8::Array::new(s, count);
-
-      for i in 0..count {
-        let frame = stack_trace
-          .get_frame(s, i as usize)
-          .expect("No frame found");
-        let frame_obj = v8::Object::new(s);
-        frames.set(context, v8::Integer::new(s, i).into(), frame_obj.into());
-        frame_obj.set(
-          context,
-          v8::String::new(s, "line").unwrap().into(),
-          v8::Integer::new(s, frame.get_line_number() as i32).into(),
-        );
-        frame_obj.set(
-          context,
-          v8::String::new(s, "column").unwrap().into(),
-          v8::Integer::new(s, frame.get_column() as i32).into(),
-        );
-
-        if let Some(function_name) = frame.get_function_name(s) {
-          frame_obj.set(
-            context,
-            v8::String::new(s, "functionName").unwrap().into(),
-            function_name.into(),
-          );
-        }
-
-        let script_name = match frame.get_script_name_or_source_url(s) {
-          Some(name) => name,
-          None => v8::String::new(s, "<unknown>").unwrap(),
-        };
-        frame_obj.set(
-          context,
-          v8::String::new(s, "scriptName").unwrap().into(),
-          script_name.into(),
-        );
-
-        frame_obj.set(
-          context,
-          v8::String::new(s, "isEval").unwrap().into(),
-          v8::Boolean::new(s, frame.is_eval()).into(),
-        );
-
-        frame_obj.set(
-          context,
-          v8::String::new(s, "isConstructor").unwrap().into(),
-          v8::Boolean::new(s, frame.is_constructor()).into(),
-        );
-
-        frame_obj.set(
-          context,
-          v8::String::new(s, "isWasm").unwrap().into(),
-          v8::Boolean::new(s, frame.is_wasm()).into(),
-        );
-      }
-
-      frames
-    } else {
-      // No stack trace. We only have one stack frame of info..
-      let frames = v8::Array::new(s, 1);
-      let frame_obj = v8::Object::new(s);
-      frames.set(context, v8::Integer::new(s, 0).into(), frame_obj.into());
-
-      frame_obj.set(
-        context,
-        v8::String::new(s, "scriptResourceName").unwrap().into(),
-        script_resource_name,
-      );
-      frame_obj.set(
-        context,
-        v8::String::new(s, "line").unwrap().into(),
-        v8::Integer::new(s, line_number as i32).into(),
-      );
-      frame_obj.set(
-        context,
-        v8::String::new(s, "column").unwrap().into(),
-        v8::Integer::new(s, message.get_start_column() as i32).into(),
-      );
-
-      frames
-    };
-
-    json_obj.set(
-      context,
-      v8::String::new(s, "frames").unwrap().into(),
-      frames.into(),
-    );
-
-    json_obj
-  }
-}
-
-fn script_origin<'a>(
+pub fn script_origin<'a>(
   s: &mut impl v8::ToLocal<'a>,
   resource_name: v8::Local<'a, v8::String>,
 ) -> v8::ScriptOrigin<'a> {
@@ -510,7 +58,7 @@ fn script_origin<'a>(
   )
 }
 
-fn module_origin<'a>(
+pub fn module_origin<'a>(
   s: &mut impl v8::ToLocal<'a>,
   resource_name: v8::Local<'a, v8::String>,
 ) -> v8::ScriptOrigin<'a> {
@@ -673,7 +221,7 @@ pub type Snapshot1 = v8::OwnedStartupData;
 pub type Snapshot2<'a> = v8::StartupData<'a>;
 
 #[allow(non_camel_case_types)]
-type deno_recv_cb = unsafe fn(
+pub type deno_recv_cb = unsafe fn(
   user_data: *mut c_void,
   op_id: OpId,
   control_buf: deno_buf,
@@ -684,7 +232,7 @@ type deno_recv_cb = unsafe fn(
 /// Embedder must call deno_dyn_import_done() with the specified id and
 /// the module.
 #[allow(non_camel_case_types)]
-type deno_dyn_import_cb = fn(
+pub type deno_dyn_import_cb = fn(
   user_data: *mut c_void,
   specifier: &str,
   referrer: &str,
@@ -698,7 +246,7 @@ pub type deno_mod = i32;
 pub type deno_dyn_import_id = i32;
 
 #[allow(non_camel_case_types)]
-type deno_resolve_cb = unsafe extern "C" fn(
+pub type deno_resolve_cb = unsafe extern "C" fn(
   resolve_context: *mut c_void,
   specifier: *const c_char,
   referrer: deno_mod,
@@ -731,15 +279,6 @@ impl Deref for SnapshotConfig {
   }
 }
 
-#[repr(C)]
-pub struct deno_config {
-  pub will_snapshot: c_int,
-  pub load_snapshot: Option<SnapshotConfig>,
-  pub shared: deno_buf,
-  pub recv_cb: deno_recv_cb,
-  pub dyn_import_cb: deno_dyn_import_cb,
-}
-
 pub unsafe fn deno_init() {
   let platform = v8::platform::new_default_platform();
   v8::V8::initialize_platform(platform);
@@ -757,7 +296,7 @@ pub unsafe fn deno_init() {
 }
 
 lazy_static! {
-  static ref EXTERNAL_REFERENCES: v8::ExternalReferences =
+  pub static ref EXTERNAL_REFERENCES: v8::ExternalReferences =
     v8::ExternalReferences::new(&[
       v8::ExternalReference {
         function: bindings::print
@@ -786,34 +325,7 @@ lazy_static! {
     ]);
 }
 
-pub unsafe fn deno_new_snapshotter(config: deno_config) -> *mut DenoIsolate {
-  assert_ne!(config.will_snapshot, 0);
-  // TODO(ry) Support loading snapshots before snapshotting.
-  assert!(config.load_snapshot.is_none());
-  let mut creator = v8::SnapshotCreator::new(Some(&EXTERNAL_REFERENCES));
-
-  let mut d = Box::new(DenoIsolate::new(config));
-  let isolate = creator.get_owned_isolate();
-
-  let mut locker = v8::Locker::new(&isolate);
-  {
-    let mut hs = v8::HandleScope::new(&mut locker);
-    let scope = hs.enter();
-    let mut context = v8::Context::new(scope);
-    // context.enter();
-    d.context_.set(scope, context);
-    creator.set_default_context(context);
-    initialize_context(scope, context);
-    // context.exit();
-  }
-  d.add_isolate(isolate);
-
-  d.snapshot_creator_ = Some(creator);
-
-  Box::into_raw(d)
-}
-
-fn initialize_context<'a>(
+pub fn initialize_context<'a>(
   scope: &mut impl v8::ToLocal<'a>,
   mut context: v8::Local<v8::Context>,
 ) {
@@ -901,91 +413,6 @@ fn initialize_context<'a>(
   context.exit();
 }
 
-pub unsafe fn deno_new(config: deno_config) -> *mut DenoIsolate {
-  if config.will_snapshot != 0 {
-    return deno_new_snapshotter(config);
-  }
-
-  let load_snapshot_is_null = config.load_snapshot.is_none();
-
-  let mut d = Box::new(DenoIsolate::new(config));
-  let mut params = v8::Isolate::create_params();
-  params.set_array_buffer_allocator(v8::new_default_allocator());
-  params.set_external_references(&EXTERNAL_REFERENCES);
-  if let Some(ref mut snapshot) = d.snapshot_ {
-    params.set_snapshot_blob(snapshot);
-  }
-
-  let isolate = v8::Isolate::new(params);
-  d.add_isolate(isolate);
-
-  let mut locker = v8::Locker::new(d.isolate_.as_ref().unwrap());
-  {
-    let mut hs = v8::HandleScope::new(&mut locker);
-    let scope = hs.enter();
-    let mut context = v8::Context::new(scope);
-
-    if load_snapshot_is_null {
-      // If no snapshot is provided, we initialize the context with empty
-      // main source code and source maps.
-      initialize_context(scope, context);
-    }
-    d.context_.set(scope, context);
-  }
-  Box::into_raw(d)
-}
-
-pub unsafe fn deno_check_promise_errors(i_mut: &mut DenoIsolate) {
-  /*
-  if (d->pending_promise_map_.size() > 0) {
-    auto* isolate = d->isolate_;
-    v8::Locker locker(isolate);
-    v8::Isolate::Scope isolate_scope(isolate);
-    v8::HandleScope handle_scope(isolate);
-    auto context = d->context_.Get(d->isolate_);
-    v8::Context::Scope context_scope(context);
-
-    auto it = d->pending_promise_map_.begin();
-    while (it != d->pending_promise_map_.end()) {
-      auto error = it->second.Get(isolate);
-      deno::HandleException(context, error);
-      it = d->pending_promise_map_.erase(it);
-    }
-  }
-  */
-  let isolate = i_mut.isolate_.as_ref().unwrap();
-
-  if i_mut.pending_promise_map_.is_empty() {
-    return;
-  }
-
-  let mut locker = v8::Locker::new(isolate);
-  assert!(!i_mut.context_.is_empty());
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  let mut context = i_mut.context_.get(scope).unwrap();
-  context.enter();
-
-  let pending_promises: Vec<(i32, v8::Global<v8::Value>)> =
-    i_mut.pending_promise_map_.drain().collect();
-  for (promise_id, mut handle) in pending_promises {
-    let error = handle.get(scope).expect("Empty error handle");
-    i_mut.handle_exception(scope, context, error);
-    handle.reset(scope);
-  }
-
-  context.exit();
-}
-
-pub unsafe fn deno_throw_exception(i_mut: &mut DenoIsolate, text: &str) {
-  let isolate = i_mut.isolate_.as_ref().unwrap();
-  let mut locker = v8::Locker::new(isolate);
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  let msg = v8::String::new(scope, text).unwrap();
-  isolate.throw_exception(msg.into());
-}
-
 pub unsafe fn deno_import_buf<'sc>(
   scope: &mut impl v8::ToLocal<'sc>,
   buf: deno_buf,
@@ -1041,160 +468,7 @@ pub unsafe fn deno_import_buf<'sc>(
   v8::Uint8Array::new(ab, 0, buf.data_len).expect("Failed to create UintArray8")
 }
 
-pub unsafe fn deno_respond(
-  deno_isolate: &mut DenoIsolate,
-  core_isolate: *const c_void,
-  op_id: OpId,
-  buf: deno_buf,
-) {
-  /*
-  auto* d = deno::unwrap(d_);
-  if (d->current_args_ != nullptr) {
-    // Synchronous response.
-    // Note op_id is not passed back in the case of synchronous response.
-    if (buf.data_ptr != nullptr && buf.data_len > 0) {
-      auto ab = deno::ImportBuf(d, buf);
-      d->current_args_->GetReturnValue().Set(ab);
-    }
-    d->current_args_ = nullptr;
-    return;
-  }
-  */
-
-  if !deno_isolate.current_args_.is_null() {
-    // Synchronous response.
-    // Note op_id is not passed back in the case of synchronous response.
-    if !buf.data_ptr.is_null() && buf.data_len > 0 {
-      let isolate = deno_isolate.isolate_.as_ref().unwrap();
-      let mut locker = v8::Locker::new(isolate);
-      assert!(!deno_isolate.context_.is_empty());
-      let mut hs = v8::HandleScope::new(&mut locker);
-      let scope = hs.enter();
-      let ab = deno_import_buf(scope, buf);
-      let info: &v8::FunctionCallbackInfo =
-        unsafe { &*deno_isolate.current_args_ };
-      let rv = &mut info.get_return_value();
-      rv.set(ab.into())
-    }
-    deno_isolate.current_args_ = std::ptr::null();
-    return;
-  }
-
-  /*
-  // Asynchronous response.
-  deno::UserDataScope user_data_scope(d, user_data);
-  v8::Isolate::Scope isolate_scope(d->isolate_);
-  v8::HandleScope handle_scope(d->isolate_);
-
-  auto context = d->context_.Get(d->isolate_);
-  v8::Context::Scope context_scope(context);
-
-  v8::TryCatch try_catch(d->isolate_);
-
-  auto recv_ = d->recv_.Get(d->isolate_);
-  if (recv_.IsEmpty()) {
-    d->last_exception_ = "Deno.core.recv has not been called.";
-    return;
-  }
-
-  v8::Local<v8::Value> args[2];
-  int argc = 0;
-
-  if (buf.data_ptr != nullptr) {
-    args[0] = v8::Integer::New(d->isolate_, op_id);
-    args[1] = deno::ImportBuf(d, buf);
-    argc = 2;
-  }
-
-  auto v = recv_->Call(context, context->Global(), argc, args);
-
-  if (try_catch.HasCaught()) {
-    CHECK(v.IsEmpty());
-    deno::HandleException(context, try_catch.Exception());
-  }
-  */
-
-  let core_isolate: *mut c_void = unsafe { std::mem::transmute(core_isolate) };
-  deno_isolate.core_isolate_ = core_isolate;
-
-  let isolate = deno_isolate.isolate_.as_ref().unwrap();
-  // println!("deno_execute -> Isolate ptr {:?}", isolate);
-  let mut locker = v8::Locker::new(isolate);
-  assert!(!deno_isolate.context_.is_empty());
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  let mut context = deno_isolate.context_.get(scope).unwrap();
-  context.enter();
-
-  let mut try_catch = v8::TryCatch::new(scope);
-  let tc = try_catch.enter();
-
-  let recv_ = deno_isolate.recv_.get(scope);
-
-  if recv_.is_none() {
-    let msg = "Deno.core.recv has not been called.".to_string();
-    deno_isolate.last_exception_ = Some(msg);
-    return;
-  }
-
-  let mut argc = 0;
-  let mut args: Vec<v8::Local<v8::Value>> = vec![];
-
-  if !buf.data_ptr.is_null() {
-    argc = 2;
-    let op_id = v8::Integer::new(scope, op_id as i32);
-    args.push(op_id.into());
-    let buf = deno_import_buf(scope, buf);
-    args.push(buf.into());
-  }
-
-  let global = context.global(scope);
-  let maybe_value =
-    recv_
-      .unwrap()
-      .call(scope, context, global.into(), argc, args);
-
-  if tc.has_caught() {
-    assert!(maybe_value.is_none());
-    deno_isolate.handle_exception(scope, context, tc.exception().unwrap());
-  }
-  context.exit();
-  deno_isolate.core_isolate_ = std::ptr::null_mut();
-}
-
-pub unsafe fn deno_execute(
-  i_mut: &mut DenoIsolate,
-  core_isolate: *mut c_void,
-  js_filename: &str,
-  js_source: &str,
-) {
-  i_mut.core_isolate_ = core_isolate;
-  let isolate = i_mut.isolate_.as_ref().unwrap();
-  // println!("deno_execute -> Isolate ptr {:?}", isolate);
-  let mut locker = v8::Locker::new(isolate);
-  assert!(!i_mut.context_.is_empty());
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  let mut context = i_mut.context_.get(scope).unwrap();
-  context.enter();
-
-  i_mut.execute(scope, context, js_filename, js_source);
-
-  context.exit();
-  i_mut.core_isolate_ = std::ptr::null_mut();
-  /*
-  auto* d = deno::unwrap(d_);
-  deno::UserDataScope user_data_scope(d, user_data);
-  auto* isolate = d->isolate_;
-  v8::Locker locker(isolate);
-  v8::Isolate::Scope isolate_scope(isolate);
-  v8::HandleScope handle_scope(isolate);
-  auto context = d->context_.Get(d->isolate_);
-  CHECK(!context.IsEmpty());
-  execute(context, js_filename, js_source);
-  */
-}
-
+/*
 pub unsafe fn deno_terminate_execution(i: *mut DenoIsolate) {
   /*
   deno::DenoIsolate* d = reinterpret_cast<deno::DenoIsolate*>(d_);
@@ -1224,219 +498,6 @@ pub unsafe fn deno_run_microtasks(i: *mut isolate, core_isolate: *mut c_void) {
   deno_isolate.core_isolate_ = std::ptr::null_mut();
 }
 
-// Modules
-
-fn resolve_callback(
-  context: v8::Local<v8::Context>,
-  specifier: v8::Local<v8::String>,
-  referrer: v8::Local<v8::Module>,
-) -> *mut v8::Module {
-  let mut cbs = v8::CallbackScope::new(context);
-  let cb_scope = cbs.enter();
-  let isolate = cb_scope.isolate();
-  let deno_isolate: &mut DenoIsolate =
-    unsafe { &mut *(isolate.get_data(0) as *mut DenoIsolate) };
-
-  let mut locker = v8::Locker::new(isolate);
-  let mut hs = v8::EscapableHandleScope::new(&mut locker);
-  let scope = hs.enter();
-
-  let referrer_id = referrer.get_identity_hash();
-  let referrer_info = deno_isolate
-    .get_module_info(referrer_id)
-    .expect("ModuleInfo not found");
-  let len_ = referrer.get_module_requests_length();
-
-  let specifier_str = specifier.to_rust_string_lossy(scope);
-
-  for i in 0..len_ {
-    let req = referrer.get_module_request(i);
-    let req_str = req.to_rust_string_lossy(scope);
-
-    if req_str == specifier_str {
-      let resolve_cb = deno_isolate.resolve_cb_.unwrap();
-      let c_str = CString::new(req_str.to_string()).unwrap();
-      let c_req_str: *const c_char = c_str.as_ptr() as *const c_char;
-      let id = unsafe {
-        resolve_cb(deno_isolate.resolve_context_, c_req_str, referrer_id)
-      };
-      let maybe_info = deno_isolate.get_module_info(id);
-
-      if maybe_info.is_none() {
-        let msg = format!(
-          "Cannot resolve module \"{}\" from \"{}\"",
-          req_str, referrer_info.name
-        );
-        let msg = v8::String::new(scope, &msg).unwrap();
-        isolate.throw_exception(msg.into());
-        break;
-      }
-
-      let child_mod =
-        maybe_info.unwrap().handle.get(scope).expect("Empty handle");
-      return &mut *scope.escape(child_mod);
-    }
-  }
-
-  std::ptr::null_mut()
-}
-
-pub unsafe fn deno_mod_instantiate(
-  i_mut: &mut DenoIsolate,
-  resolve_context: *mut c_void,
-  id: deno_mod,
-  resolve_cb: deno_resolve_cb,
-) {
-  i_mut.resolve_context_ = resolve_context;
-  let isolate = i_mut.isolate_.as_ref().unwrap();
-  let mut locker = v8::Locker::new(isolate);
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  assert!(!i_mut.context_.is_empty());
-  let mut context = i_mut.context_.get(scope).unwrap();
-  context.enter();
-  let mut try_catch = v8::TryCatch::new(scope);
-  let tc = try_catch.enter();
-
-  assert!(i_mut.resolve_cb_.is_none());
-  i_mut.resolve_cb_ = Some(resolve_cb);
-
-  let maybe_info = i_mut.get_module_info(id);
-
-  if maybe_info.is_none() {
-    return;
-  }
-
-  let module_handle = &maybe_info.unwrap().handle;
-  let mut module = module_handle.get(scope).unwrap();
-
-  if module.get_status() == v8::ModuleStatus::Errored {
-    return;
-  }
-
-  let maybe_ok = module.instantiate_module(context, resolve_callback);
-  assert!(maybe_ok.is_some() || tc.has_caught());
-  i_mut.resolve_cb_.take();
-
-  if tc.has_caught() {
-    i_mut.handle_exception(scope, context, tc.exception().unwrap());
-  }
-
-  context.exit();
-  i_mut.resolve_context_ = std::ptr::null_mut();
-}
-
-pub unsafe fn deno_mod_evaluate(
-  deno_isolate: &mut DenoIsolate,
-  core_isolate: *const c_void,
-  id: deno_mod,
-) {
-  let core_isolate: *mut c_void = unsafe { std::mem::transmute(core_isolate) };
-  deno_isolate.core_isolate_ = core_isolate;
-  let isolate = deno_isolate.isolate_.as_ref().unwrap();
-  let mut locker = v8::Locker::new(isolate);
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  assert!(!deno_isolate.context_.is_empty());
-  let mut context = deno_isolate.context_.get(scope).unwrap();
-  context.enter();
-
-  let info = deno_isolate
-    .get_module_info(id)
-    .expect("ModuleInfo not found");
-  let mut module = info.handle.get(scope).expect("Empty module handle");
-  let mut status = module.get_status();
-
-  if status == v8::ModuleStatus::Instantiated {
-    let ok = module.evaluate(scope, context).is_some();
-    // Update status after evaluating.
-    status = module.get_status();
-    if ok {
-      assert!(
-        status == v8::ModuleStatus::Evaluated
-          || status == v8::ModuleStatus::Errored
-      );
-    } else {
-      assert!(status == v8::ModuleStatus::Errored);
-    }
-  }
-
-  match status {
-    v8::ModuleStatus::Evaluated => {
-      deno_isolate.last_exception_handle_.reset(scope);
-      deno_isolate.last_exception_.take();
-    }
-    v8::ModuleStatus::Errored => {
-      deno_isolate.handle_exception(scope, context, module.get_exception());
-    }
-    other => panic!("Unexpected module status {:?}", other),
-  };
-
-  context.exit();
-  deno_isolate.core_isolate_ = std::ptr::null_mut();
-}
-
-/// Call exactly once for every deno_dyn_import_cb.
-pub unsafe fn deno_dyn_import_done(
-  deno_isolate: &mut DenoIsolate,
-  core_isolate: *const c_void,
-  id: deno_dyn_import_id,
-  mod_id: deno_mod,
-  error_str: Option<String>,
-) {
-  assert!(
-    (mod_id == 0 && error_str.is_some())
-      || (mod_id != 0 && error_str.is_none())
-      || (mod_id == 0 && !deno_isolate.last_exception_handle_.is_empty())
-  );
-
-  let core_isolate: *mut c_void = unsafe { std::mem::transmute(core_isolate) };
-  deno_isolate.core_isolate_ = core_isolate;
-
-  let isolate = deno_isolate.isolate_.as_ref().unwrap();
-  let mut locker = v8::Locker::new(isolate);
-  let mut hs = v8::HandleScope::new(&mut locker);
-  let scope = hs.enter();
-  assert!(!deno_isolate.context_.is_empty());
-  let mut context = deno_isolate.context_.get(scope).unwrap();
-  context.enter();
-
-  // TODO(ry) error on bad import_id.
-  let mut resolver_handle = deno_isolate.dyn_import_map_.remove(&id).unwrap();
-  /// Resolve.
-  let mut resolver = resolver_handle.get(scope).unwrap();
-  resolver_handle.reset(scope);
-
-  let maybe_info = deno_isolate.get_module_info(mod_id);
-
-  if let Some(info) = maybe_info {
-    // Resolution success
-    let mut module = info.handle.get(scope).unwrap();
-    assert_eq!(module.get_status(), v8::ModuleStatus::Evaluated);
-    let module_namespace = module.get_module_namespace();
-    resolver.resolve(context, module_namespace).unwrap();
-  } else {
-    // Resolution error.
-    if let Some(error_str) = error_str {
-      let msg = v8::String::new(scope, &error_str).unwrap();
-      let isolate = context.get_isolate();
-      isolate.enter();
-      let e = v8::type_error(scope, msg);
-      isolate.exit();
-      resolver.reject(context, e).unwrap();
-    } else {
-      let e = deno_isolate.last_exception_handle_.get(scope).unwrap();
-      deno_isolate.last_exception_handle_.reset(scope);
-      deno_isolate.last_exception_.take();
-      resolver.reject(context, e).unwrap();
-    }
-  }
-
-  isolate.run_microtasks();
-
-  context.exit();
-  deno_isolate.core_isolate_ = std::ptr::null_mut();
-}
 
 pub fn deno_snapshot_new(
   deno_isolate: &mut DenoIsolate,
@@ -1463,3 +524,4 @@ pub fn deno_snapshot_new(
 pub unsafe fn deno_snapshot_delete(s: &mut deno_snapshot) {
   todo!()
 }
+*/
